Sprite's scripts list
Functions added:
public void createNewScript(Sprite sprite)
public Script[] getScriptList()
public SpriteName getSpriteName()

Functions changed:
public Script getScript(int scriptNumber) - it now receibes the
scriptNumber it wants, although this function might seem redundant
because there is a new function that returns the whole list. It is kept
for just-in-case purposes.

----------

Also a sprite now has a structure* of several** scripts. NOTE: It hasn't
been tested like this but it now has the option to do it.
To create a new script you have to call the funcion createNewScript that
receibes a Sprite object and adds a new script to that sprite. Example:

createNewScript(Sprite sp);

whenever this function is named, the user has to access the list of
scripts of that sprite so it can access the new script.

*structure: List.
** several: more than one.